Conflicts:
	SVGKit-iOS.xcodeproj/project.pbxproj
	Source/DOM classes/SVG-DOM/SVGGElement.m
	Source/DOM classes/Unported or Partial DOM/SVGElement.m
	Source/DOM classes/Unported or Partial DOM/SVGGradientElement.m
	Source/DOM classes/Unported or Partial DOM/SVGSVGElement.m
	Source/DOM classes/Unported or Partial DOM/SVGTextElement.m
	Source/Parsers/SVGKParser.m
	Source/SVGKImage.m
	Source/SVGKit.h
	XCodeProjectData/SVGKit-iOS/SVGKit-iOS-Prefix.pch

+++ Source/DOM classes/Core DOM/NamedNodeMap.m
@@ -45,7 +45,7 @@
 		 
 		 NB: according to spec, this behaviour is:
 		 
-		    "The result depends on the implementation"
+		 "The result depends on the implementation"
 		 
 		 I've chosen to implement it the most user-friendly way possible. It is NOT the best
 		 solution IMHO - the spec authors should have defined the outcome!
@@ -86,7 +86,7 @@
 
 -(unsigned long)length
 {
-	int count = [self.internalDictionary count];
+	NSUInteger count = [self.internalDictionary count];
 	
 	for( NSDictionary* namespaceDict in self.internalDictionaryOfNamespaces )
 	{
